refactor(react-sdk): remove legacy theme system for cleaner codebase
- Remove all legacy theme files (ChatWidget.compat.ts, ChatWidget.styles.legacy.ts)
- Remove customTheme prop and backward compatibility code
- Update documentation to reflect minimal 4-property theme system
- Remove unused typedoc-plugin-markdown dependency
- Clean up all legacy theme exports and imports

BREAKING CHANGE: Removed customTheme prop and legacy theme support. Use minimalTheme or CSS custom properties instead.
