+++ packages/react-sdk/src/components/index.ts
@@ -1,4 +1,2 @@
 export { ChatWidget } from './ChatWidget.js';
-export type { ChatWidgetProps } from './ChatWidget.js';
-export type { ThemeColors } from './ChatWidget.styles.legacy.js';
-export { lightTheme, darkTheme } from './ChatWidget.styles.legacy.js';
+export type { ChatWidgetProps, MinimalTheme } from './ChatWidget.js';
